use size_t casting to prevent allocation failure
- the cast to size_t is done before multiplying with size_of
- size_of now returns a size_t
- This prevent reduction of capacity in 32bit mode when cgsize_t is an int.

+++ src/cgns_header.h
@@ -1248,7 +1248,7 @@ int cgi_zone_no(cgns_base *base, char *zonename, int *zone_no);
 
 /* miscellaneous */
 int cgi_sort_names(int n, double *ids);
-int size_of(const char_33 adf_type);
+size_t size_of(const char_33 adf_type);
 char *type_of(char_33 data_type);
 int cgi_check_strlen(char const * string);
 int cgi_check_strlen_x2(char const *string);
